+++ CHANGELOG.md
@@ -5,9 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.7.0] - 2022-08-10
+### Added
+- Env variable `HTTP_CHECK_PERIOD` to specify period of HTTP checks. Default 5 sec
+- Added prometheus metrics. At `/metrics` location
+    - DNSLookup time (Histogram)
+    - TCP Connection time (Histogram)
+    - ServerProcess time (Histogram)
+    - Content transfer time (Histogram)
+    - Response Codes From Hosts (Counter)
+
 ## [1.6.0] - 2022-08-10
 ### Changed
-- HOSTS env var renamed to  HTTP_HOSTS
+- HOSTS env var renamed to  `HTTP_HOSTS`
 
 ## [1.5.0] - 2022-08-10
 ### Changed